# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Faarooq Backend API",
    "version": "0.1.0",
    "description": "OpenAPI spec for the Twitter-like backend (Express + Supabase)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local" },
    { "url": "https://faarooq-backend.onrender.com", "description": "Production" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    },
    "responses": {
      "RateLimited": {
        "description": "Too many requests",
        "headers": {
          "Retry-After": { "schema": { "type": "integer" }, "description": "Seconds until reset" },
          "X-RateLimit-Limit": { "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "schema": { "type": "integer" } },
          "X-RateLimit-Reset": {
            "schema": { "type": "integer" },
            "description": "Unix epoch seconds when window resets"
          }
        },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" },
              "details": { "type": "object" }
            },
            "required": ["code"]
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "paths": {
...
    "/api/notifications": {
      "get": {
        "summary": "List notifications",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "cursor", "in": "query", "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "OK" },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/api/moderation/block": {
      "post": {
        "summary": "Block a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "user_id": { "type": "string", "format": "uuid" } },
                "required": ["user_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/moderation/unblock": {
      "post": {
        "summary": "Unblock a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "user_id": { "type": "string", "format": "uuid" } },
                "required": ["user_id"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/moderation/mute": {
      "post": {
        "summary": "Mute a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "user_id": { "type": "string", "format": "uuid" } },
                "required": ["user_id"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/moderation/unmute": {
      "post": {
        "summary": "Unmute a user",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "user_id": { "type": "string", "format": "uuid" } },
                "required": ["user_id"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/moderation/report": {
      "post": {
        "summary": "Report a user or post",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": { "type": "string", "format": "uuid" },
                  "post_id": { "type": "string", "format": "uuid" },
                  "reason": { "type": "string" }
                },
                "required": ["reason"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": {
            "description": "Invalid payload",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/notifications/read": {
      "post": {
        "summary": "Mark notifications read",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": { "type": "array", "items": { "type": "string", "format": "uuid" } }
                },
                "required": []
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  }
}